requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cleanup==8.0.0
django-cuser==2017.3.16
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.18.11
django-tinymce==3.6.1
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
